set up this repo to use Transcrypt
Transcrypt is a library which can compile Python to Javascript.
By using this, the common code can be written and maintained once in Python. see README

Also added an __init__.py file which is used as the entry point for Transcrypt, and will also be the entry point if we distribute this library as a package on pip or npm