patch: 0.1.3
Commented out all uses of the @memoize decorator.

I thought this decorator would improve performance, but it is actually causing significant slowness when memoized functions are called in JS. Maybe something to do with how **kwargs is handled when converting to JS.

May reinstate later with a JS specific implementation.