implement "SurveysSummary' module
This module allows us to create a 'summary' of how many survey responses have been completed vs. prompted, given a list of trips.
get_surveys_summary returns a dict in the form of:
```
{
  'SurveyNameA': {
    'answered': 4,
    'unanswered': 3,
    'mismatched': 0
  },
  'SuveryNameB': {
    ...
    ...
  }
}
```
These totals represent the number of trips for which a survey was answered / unanswered / mismatched. "Mismatched" means the prompted survey and answered survey are not the same (this could happen if the config was updated causing a different survey to be assigned to the trip).

Calculating this requires us to determine which surveys are prompted for which trips, using the 'showsIf' expressions in the app config. This is handled in conditional_surveys.py, which is basically a Python implementation of conditionalSurveys.ts (added in https://github.com/e-mission/e-mission-phone/pull/1129)
Additional discussion around evaluating the expressions in Python: https://github.com/e-mission/e-mission-docs/issues/1051#issuecomment-2033320906