reorganize folder structure
For e-mission-common to function as both a JS and Python package (compatible with npm and pip respectively), there need to be a few changes.
First, the directory that contains __init__.js needs to be the name of the Python package. We will call it `emcommon` and rename the directory accordingly.

Second, it's simpler and easier to just have the npm package at the same level as the Python package. So instead of tucking it away under the /js directory, we can keep package.json at the root level. The compiled JS files can go under a directory called `emcommon_translated_js`.

Moved all these files, updated imports, updated compile_to_js.sh, updated README.