add functions to surveys_summary.py
`get_response_rate_rankings` returns a dict of user_ids to users' overall survey response rate. This will be needed to implement a leaderboard on the app Dashboard tab.

`get_median_response_rates` would return the median response rate for each of the surveys, by calculating each user's response % and then finding the median user.
Commented this out for now because I think we may end up just using overall averages (ie total number of answered surveys / total number of surveys, across all users)
It depends what we actually want to compare. Should it be the user response % vs. the response % of the group as a whole, or should it be the user response % vs. the response % of a "typical user" (ie median user)