ble_matching: access dicts by []
Reminder to myself: Code in e-mission-common should use bracket notation since we're often working with inputs that could be either JS objects or Python dicts. Dot notation will not work on Python dicts!

Also, a fix on line 78. `bluetooth_major_minor` is defined in the config as an *array* of possible major:minor strings, not a single string.