reflect the new structure of LabelMap on the phone
Since 02655ba28913a06caa80795aa3a07a4b15d06ef3 on e-mission-phone, 'SURVEY' is not used anymore; the name of the survey is used. We don't know which survey we're looking for, but if the trip is present in the LabelMap it will have exactly 1 survey. So we can grab the first value and get the name of the survey from there.