0.5.2: patch to filter 'untracked' composite_trips correctly
Flattening and filtering in one step caused a key error because 'origin_key' may not be present at the root level until flattening has been done.
So it is necessary to flatten first, then filter.

Also updated variable names and comments to reflect that the list of trips passed into generate_summaries can be either composite or confirmed trips. After filtering, we can treat them all as confirmed trips, so the types of the other functions' arguments should be named as confirmed trips.