expand base modes to include "met" and "footprint"
Up to now, we've had base modes only for the purpose of having a designated color and icon for any rich modes that get defined.
All of the MET and carbon footprint info was defined for each and every rich mode (in the label-options JSON)

Now, base modes will have a base "met" and "footprint".
The mets are copied over from what was in e-mission-phone (reformatted to Python).
Rich modes previously had a 'kgCo2PerKm' value. As part of the effort to de-couple energy calculations from carbon calculations (https://github.com/e-mission/e-mission-docs/issues/954), this is replaced by a 'footprint' field, which has energy intensities broken down by fuel type. For transit, we describe the mapping to mode(s) in the NTD.
Rich modes can still override "footprint" or "met", or any of these properties.

It will no longer be necessary to include "met_equivalent" in rich modes. If "met" is not present, it will be implied as being the met of the base mode.

Now our "base modes" and "rich modes" line up better and we can describe it as an inheritance pattern.