ntd script: output
similar to 541fdbfd8156ca66d6a52528e70a1a1812da9b91, I am changing the outputs to be organized as .json files by year. This makes it easier to track what URLs the data is sourced from because we are including metadata for this in the output files.

This meant restructuing this script a little bit; to establish a deterministic relationship between the input URLs for a year and the output JSON files for a year, each year needs to be completely isolated from the others. The method of dropping outliers (which had relied on the combined data from all years to determine upper/lower bounds) needed to be modified to only consider the current year.