use the same miniconda installation as e-mission-server
Instead of using venv, we can use miniconda like e-mission-server does.
In fact, we can run the same exact scripts to set up the miniconda installation (if it doesn't already exist) and just set up a new conda environment within. I accomplished this by writing a script that clones the setup scripts from e-mission-server.

This way, we are guaranteed to be using the same python distribution while testing e-mission-common as we are running on e-mission-server.