implement unified testing approach
(see README)
Enables a way for tests to be written once in python. They can be run in pytest (with bin/run_pytest) or in jest (with bin/run_jest, which compiles the py test suites into JS, then jest runs on the output JS)

Add tests for base_modes and emcommon.util as the first examples of how this will work.

The existing tests that already have separate .js files can be kept how they are.